Добавление панели админа

+++ client/js/users.js
@@ -73,18 +73,23 @@ var main = function (UsersObjects) {
 		if ($input.val() !== "") {
 			if ($input.val() !== null && $input.val().trim() !== "") {
 				var username = $input.val();
-				if (confirm("Вы уверены, что хотете удалить пользователя " + username + "?")) {
-					$.ajax({
-						'url': '/users/'+username,
-						'type': 'DELETE',
-					}).done(function(responde) {
-						console.log(responde);
-						$input.val("");
-						alert('Пользователь успешно удален');
-					}).fail(function(jqXHR, textStatus, error) {
-						console.log(error);
-						alert("Произошла ошибка!\n"+jqXHR.status + " " + jqXHR.textStatus);	
-					});
+				if (username == "admin") {
+					alert('Нельзя удалить администратора!!!');
+				}
+				else {
+					if (confirm("Вы уверены, что хотете удалить пользователя " + username + "?")) {
+						$.ajax({
+							'url': '/users/'+username,
+							'type': 'DELETE',
+						}).done(function(responde) {
+							console.log(responde);
+							$input.val("");
+							alert('Пользователь успешно удален');
+						}).fail(function(jqXHR, textStatus, error) {
+							console.log(error);
+							alert("Произошла ошибка!\n"+jqXHR.status + " " + jqXHR.textStatus);	
+						});
+					}
 				}
 			}
 		}
